Trigger sepolicy rebuild on certificate changes
Add *.pem files to the prerequisites for sepolicy, meaning that changes
to such files properly trigger a regeneration of mac_permissions.

Prior to this, when a certificate file has changed, this will not be
detected unless a clean build is performed or unless the file that
references the certificate, i.e. keys.conf, changes (mtime-based).

Change-Id: I88bbf63e0ec8893dfe7fa48dc4aab312b5f749eb